Separate functionality specific to handling Azure blobs into a separate module.
