+++ src/lib/phdi-building-blocks/phdi_building_blocks/fhir.py
@@ -11,38 +11,11 @@ from azure.identity import DefaultAzureCredential
 from azure.storage.blob import ContainerClient
 
 
-def get_blob_client(container_url: str) -> ContainerClient:
-    """Use whatever creds Azure can find to authenticate with the storage container"""
-    creds = DefaultAzureCredential()
-    return ContainerClient.from_container_url(container_url, credential=creds)
-
-
 def generate_filename(blob_name: str, message_index: int) -> str:
     """Strip the file type suffix from the blob name, and append message index."""
     fname = blob_name.split("/")[-1]
     fname, _ = fname.rsplit(".", 1)
     return f"{fname}-{message_index}"
-
-
-def store_data(
-    container_url: str,
-    prefix: str,
-    filename: str,
-    bundle_type: str,
-    message_json: dict = None,
-    message: str = None,
-) -> None:
-    """
-    Store the given data, which is either a FHIR bundle or an HL7 message in the
-    appropriate output container
-    """
-    client = get_blob_client(container_url)
-    blob = client.get_blob_client(str(pathlib.Path(prefix) / bundle_type / filename))
-    if message_json is not None:
-        blob.upload_blob(json.dumps(message_json).encode("utf-8"), overwrite=True)
-    elif message is not None:
-        blob.upload_blob(bytes(message, "utf-8"), overwrite=True)
-
 
 class AzureFhirserverCredentialManager:
     """Manager for handling Azure credentials for access to the FHIR server"""
